Force some build dependencies.
THe extension loader imports the plugins as text strings,
but the loading process actually depends on the plugin
modules being present.

The git module is imported inline, which is caught in
most cases by buildgen, but it wasn't in the ast,
presumably because the if block wasn't triggered.

The comments on the BUILD file forces buildgen to allow
the dep.