Do not traverse the dependees of regexp excluded targets in ChangedCalculator.
Suppose that A depends on B.  If B's source is changed and we call
`./pants test-changed --include-dependees=direct`, then both
B and A will be tested.  If we exclude B using `exclude_target_regexp`,
A is still tested even though the only path to it via changed targets
is through B.

In all circumstances where a user excludes a target specifically for
test-changed, they probably intend to also exclude dependees that
would have been brought in by the excluded target.  This patch
changes the behavior of ChangeCalculator to match that
expectation.