Add webpack to contrib
This is essentially a pure copy-paste from the most
recent push to Fsq.io. I updated the imports and
I did my best to fix up the BUILD files, in case
someone decides to try and publish a contrib module.

Of special note, we uninstall node.test and node.resolve.
I have a fix for noide.test, but node.resolve is unneeded
as far as we are concerned, we replaced it with webpack.

Feel free to uncomment and experiment as needed.