Have webpack put on the compile_classpath
This is worriesome. We use a different resolver (pom-resolve) but
requireing the 'resolve' before codegen is an error in upstream Pants.

There is a task cycle when using the 'runtime_classpath'. It is unknown how
upstream Pants will handle this. Internally, we needed to put those files on the
'runtime_classpath' in order for them to be available within
resources.jar.

You should experiment and see if this works for you. You may need to
make adjustments to work with the ivy-constructed classpath.