Rework the s3 artifact cache.
Tests incoming. I took out the pyproperties
dep, since passing it the AWS credentials (the most
natural thing to assume) caused it to crash due to
a code mistake. It hasn't been updated in 6 years, so
we are diverting all config reads to Boto, and if someone
wants an alternative config, that can land independently.