Tidy up the error handling and error catching on the credential refactor.
This commit completes the change to explicitly pass the access_key and
secret_key when setting up the client.

The s3 cache will do its best to stay out of the way should folks
want it to consume their default creds, but does support passing
in a file path to a specific creds file and using those instead.

Needs a lot of test coverage work, which is what I will do next as
I migrate our tasks to use this new cache. Our tasks have needed
refactoring in the places where they ignore the results_dir.

I will do some tests and see if the problem is our YOLO task logic
or that s3 is less eager to create needed dirs. We will see.