Fix the Go buildgen to work with go-thrift
This is a simple and dirty hack - even figuring out what this code
was doing was incredibly complicated - not a very fun library.

I tried to carry it forward into something we could land upstream,
we are going to have to do that sooner or later, this is an expensice
thing to fork long term.

Here is where that ended up:
https://github.com/mateor/pants/commits/mateo.go_buildgen

I think it works fully, but it uses alternate_target_roots
to force the targets in context and that breaks a ton of the
assumptions in the tests. so that will be a pain, but maybe
a pain we have to slog through.